move repr tests to test_formatting.py
